[version 1.10.18 bugfix] fixed panic error with address nil pointer in autoComplete. Added auto cleaner duplicates for command history and added debug var

+++ scripts/orbix.go
@@ -60,6 +60,10 @@ func OrbixLoop(panicChan chan any,
 	}()
 
 	defer func() {
+		if system.Debug {
+			return
+		}
+
 		if r := recover(); r != nil {
 			user.DeleteUserFromRunningFile(system.UserName)
 			PanicText := fmt.Sprintf("Panic recovered: %v", r)
